microservices/metadata_extraction/requirements.txt
fastapi
uvicorn
kafka-python
aiokafka
cassandra-driver
imageio
Pillow
